Update All The Things
Added confirmation of image downloading and download completion.

Added dialog confirmation before deleting images.

Added realbooru

Removed work-in-progress login information from every core. There will be a login system for sites in the future, but logging in on some sites is a complicated thing to do. This is an important feature that I intend to add as some sites limit tag count if you're not logged in!